requirements.txt
asgiref
autopep8
backports.zoneinfo
boto3
botocore
certifi
cffi
charset-normalizer
click
colorama
coreapi
coreschema
cryptography==38.0.1
distlib
Django
django-cors-headers
django-countries
django-extensions
django-extra-keywords
django-gunicorn
django-rest-swagger
django-storages
djangorestframework
djangorestframework-jwt
djangorestframework-simplejwt
dnspython
drf-yasg
filelock
filetype
gunicorn
h11
idna
inflection
itypes
Jinja2
jmespath
jsonfield
MarkupSafe
mysql-connector-python
openapi-codec
packaging
Pillow
platformdirs
# protobuf
psycopg2
psycopg2-binary
pyasn1
pycodestyle
pycparser
pycryptodome
pycryptodomex
pymongo
pyparsing
python-dateutil
python-dotenv
PyJWT
requests
rsa
ruamel.yaml
ruamel.yaml.clib
s3transfer
simplejson
six
sniffio
sqlparse
tomli
typing-extensions
tzdata
uritemplate
urllib3
virtualenv
